api_evaluator/requirements.txt
python-socketio[client]==5.10.0
requests==2.31.0
pandas==2.1.0
